Created SingleEndpointSchemaConfigurator
