+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/SchemaConfigurators/AbstractSchemaConfigurationEndpointSchemaConfigurator.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GraphQLAPI\GraphQLAPI\Services\SchemaConfigurators;
+
+abstract class AbstractSchemaConfigurationEndpointSchemaConfigurator extends AbstractEndpointSchemaConfigurator
+{
+    /**
+     * The received customPostID is already the schemaConfigurationID!
+     */
+    protected function getSchemaConfigurationID(int $customPostID): ?int
+    {
+        return $customPostID;
+    }
+}
