+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/SchemaConfigurators/EndpointSchemaConfigurator.php
@@ -7,10 +7,9 @@ namespace GraphQLAPI\GraphQLAPI\Services\SchemaConfigurators;
 use GraphQLAPI\GraphQLAPI\Registries\EndpointSchemaConfigurationExecuterRegistryInterface;
 use GraphQLAPI\GraphQLAPI\Registries\ModuleRegistryInterface;
 use GraphQLAPI\GraphQLAPI\Registries\SchemaConfigurationExecuterRegistryInterface;
-use GraphQLAPI\GraphQLAPI\Services\SchemaConfigurators\AbstractEndpointSchemaConfigurator;
 use PoP\ComponentModel\Instances\InstanceManagerInterface;
 
-class EndpointSchemaConfigurator extends AbstractEndpointSchemaConfigurator
+class EndpointSchemaConfigurator extends AbstractCustomPostEndpointSchemaConfigurator
 {
     public function __construct(
         InstanceManagerInterface $instanceManager,
